+++ CHANGES.rst
@@ -1,8 +1,17 @@
 Changelog
 =========
 
-3.0.1 (unreleased)
-------------------
+3.1 (unreleased)
+----------------
+
+- Fix Plone 5 compatibility.
+  [thet]
+
+- Extend the ``uninstall`` profile with more de-registrations.
+  [thet]
+
+- Fix the blacklist XPath selector to also match elements with the class directly set on it.
+  [thet]
 
 - Add configlet option to enable lazy loading for authenticated users.
   [hvelarde]